+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.1.1
+##  26/09/2022
+
+1. [](#new)
+    * Add Option Delay
+    * read Config (SMTP Server, user, pass ... from yaml config file)
+
 # v0.1.0
 ##  20/07/2022